move action sim charts to component
