+++ src/components/Actions/ActionSummary.vue
@@ -410,57 +410,19 @@
         </div>
       </div>
       <div v-if="!embedded" class="row q-gutter-y-lg">
-        <div
-          class="column q-gutter-md"
-          v-for="chart in chartsArr"
-          :key="chart.nodeId"
-        >
-          <gchart
-            type="LineChart"
-            :data="chart.chartData"
-            :options="chart.chartOptions"
-          />
-          <div class="row justify-center q-gutter-x-md">
-            <q-btn-toggle
-              v-model="chart.chartOptions.series"
-              action-color="primary"
-              size="xs"
-              :options="[
-                {
-                  label: 'difference',
-                  value: showDifferenceConfig,
-                },
-                {
-                  label: 'values',
-                  value: showValuesConfig,
-                },
-              ]"
-            />
-
-            <q-btn-toggle
-              v-model="chart.chartOptions.vAxis.scaleType"
-              action-color="primary"
-              size="xs"
-              :options="[
-                { label: 'linear', value: 'linear' },
-                { label: 'log', value: 'mirrorLog' },
-              ]"
-            />
-          </div>
-        </div>
+        <action-sim-charts />
       </div>
     </q-form>
   </div>
 </template>
 
 <script>
-import { firebase, firebaseApp, firebaseDb, firebaseAuth } from "boot/firebase";
+import { firebaseAuth } from "boot/firebase";
 import { mapActions, mapMutations, mapGetters, mapState } from "vuex";
 import { createHelpers, mapMultiRowFields } from "vuex-map-fields";
 import { interpret } from "xstate";
 import { actionMachine } from "src/state-machines/machine-action";
 import { formatNumber } from "src/utils/util-formatNumber";
-import { GChart } from "vue-google-charts";
 
 const { mapFields } = createHelpers({
   getterType: "uiAction/getField",
@@ -480,8 +442,9 @@ export default {
     "action-relationships": require("components/Actions/Relationships/ActionRelationships.vue")
       .default,
     "select-user": require("components/Users/SelectUser.vue").default,
+    "action-sim-charts": require("components/Charts/ActionSimCharts.vue")
+      .default,
     chat: require("components/Chats/Chat.vue").default,
-    gchart: GChart,
   },
 
   props: ["action"],
@@ -491,7 +454,6 @@ export default {
       //embedded: false, //whether this component is embedded or a full page
       actionId: null,
       saveFullResults: false,
-      chartsArr: [],
       effortCostPerHrTypeOptions: [
         {
           label: "use average effort cost per hour",
@@ -502,20 +464,7 @@ export default {
           value: "use_custom",
         },
       ],
-      showValuesConfig: {
-        0: { lineWidth: 5, visibleInLegend: true },
-        1: { lineWidth: 2, visibleInLegend: true },
-        2: { lineWidth: 2, visibleInLegend: true },
-        3: { lineWidth: 0, visibleInLegend: false },
-      },
-      showDifferenceConfig: {
-        0: { lineWidth: 0, visibleInLegend: false },
-        1: { lineWidth: 0, visibleInLegend: false },
-        2: { lineWidth: 0, visibleInLegend: false },
-        3: { lineWidth: 2, visibleInLegend: true },
-      },
       actionService: interpret(actionMachine),
-
       actionStateContext: null,
     };
   },
@@ -677,100 +626,6 @@ export default {
       this.$store.dispatch("actions/updateAction", payload);
       this.$store.commit("uiAction/setUiActionChanged", false);
     },
-
-    updateChartDataForNode(nodeId) {
-      //console.log(this.resultsOfAction.timeSPoints);
-
-      // if baseline.nodes contains the selected node then load baseline for this nde
-      if (
-        this.resultsOfAction !== undefined &&
-        this.resultsOfAction.timeSPoints !== undefined &&
-        this.resultsOfAction.timeSPoints.length
-      ) {
-        let timeSPoints = this.resultsOfAction.timeSPoints;
-        let baselineValues = this.resultsOfAction.baselineNodesValues[nodeId];
-        let ifNotDoneValues = this.resultsOfAction.ifNotDoneNodesValues[nodeId];
-        let ifDoneValues = this.resultsOfAction.ifDoneNodesValues[nodeId];
-        //if nodeId does not exist in chartsDataArr then create it
-        let chart = this.chartsArr.find((chart) => chart.nodeId == nodeId);
-        if (typeof chart == "undefined") {
-          console.log("existing chart not found");
-          let unit = (chart = {
-            nodeId: nodeId,
-            chartData: [],
-            chartOptions: {
-              title: this.getNodeName(nodeId),
-              vAxis: {
-                title: this.getNodeUnit(nodeId),
-                scaleType: "linear",
-                format: "short",
-              },
-              legend: { position: "bottom" },
-              series: this.showDifferenceConfig,
-              width: 360,
-              height: 240,
-              //explorer: {}
-            },
-          });
-          this.chartsArr.push(chart);
-        } else {
-          console.log("existing chart found");
-          chart.chartData = [];
-        }
-        if (ifDoneValues.length > 0) {
-          chart.chartData.push([
-            "time",
-            "baseline",
-            "if done",
-            "if not done",
-            "done minus not done",
-          ]);
-          for (var i = 0; i < timeSPoints.length; i++) {
-            chart.chartData.push([
-              new Date(timeSPoints[i] * 1000),
-              baselineValues[i],
-              ifDoneValues[i],
-              ifNotDoneValues[i],
-              ifDoneValues[i] - ifNotDoneValues[i],
-            ]);
-          }
-        }
-      } else {
-        this.chartsArr = [];
-      }
-    },
-    updateChartsArr() {
-      if (this.currentModel && this.uiAction && this.uiAction.impacts) {
-        /*if (!this.currentModel || typeof this.uiAction.impacts == "undefined")
-        return;*/
-        //console.log("updateChartsArr");
-        let nodesToChart = [];
-        //add impacted nodes
-        this.uiAction.impacts.forEach(function (impact) {
-          nodesToChart.push(impact.nodeId);
-        });
-        //add benefit and cost nodes
-        nodesToChart.push(this.currentModel.roleNodes.orgBenefit);
-        nodesToChart.push(this.currentModel.roleNodes.orgCost);
-        nodesToChart.push(this.currentModel.roleNodes.worldBenefit);
-        nodesToChart.push(this.currentModel.roleNodes.worldCost);
-        //nodesToChart.push(this.currentModel.roleNodes.effort);
-        //nodesToChart.push(this.currentModel.roleNodes.spending);
-
-        //load data into each node
-        nodesToChart.forEach((nodeId) => this.updateChartDataForNode(nodeId));
-      }
-    },
-    getNodeName(nodeId) {
-      const found = this.nodes.find((node) => node.id == nodeId);
-      if (found) return found.name;
-      else return nodeId;
-    },
-    getNodeUnit(nodeId) {
-      const found = this.nodes.find((node) => node.id == nodeId);
-      if (found) return found.unit;
-      else return "";
-    },
   },
 
   created() {
@@ -796,12 +651,6 @@ export default {
     this.actionService.stop();
   },
   watch: {
-    nodes: function () {
-      if (this.embedded == false) this.updateChartsArr();
-    },
-    resultsOfAction: function () {
-      if (this.embedded == false) this.updateChartsArr();
-    },
     action: {
       deep: true,
       handler(newAction, oldAction) {
